deleted blogcontroller and refactoring article and comment controller
