+++ src/Controller/ArticleController.php
@@ -1,0 +1,46 @@
+<?php
+
+
+namespace Controller;
+
+
+use Model\ArticleManager;
+use Model\CommentManager;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Error\SyntaxError;
+
+class ArticleController extends Controller
+{
+    /**
+     * @return string
+     * @throws LoaderError
+     * @throws RuntimeError
+     * @throws SyntaxError
+     */
+    public function indexAction()
+    {
+        $articles = (new ArticleManager)->findAll();
+
+        return $this->render('blog.html.twig', compact('articles'));
+    }
+
+    /**
+     * @return string
+     * @throws LoaderError
+     * @throws RuntimeError
+     * @throws SyntaxError
+     */
+    public function readAction()
+    {
+        $dataId = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
+        $articles = (new ArticleManager)->find($dataId);
+
+        $comments = (new CommentManager)->findAll($dataId);;
+
+
+        return $this->render('article.html.twig', compact('articles', 'comments'));
+
+
+    }
+}
