+++ src/Controller/CommentController.php
@@ -11,26 +11,11 @@ use Twig\Error\RuntimeError;
 use Twig\Error\SyntaxError;
 
 /**
- * Class BlogController
+ * Class CommentController
  * @package Controller
  */
-class BlogController extends Controller
-
+class CommentController extends Controller
 {
-    /**
-     * @return string
-     * @throws LoaderError
-     * @throws RuntimeError
-     * @throws SyntaxError
-     */
-    public function indexAction()
-    {
-        $articles = (new ArticleManager)->findAll();
-
-        return $this->render('blog.html.twig', compact('articles'));
-    }
-
-
     /**
      * @return string
      * @throws LoaderError
@@ -62,33 +47,4 @@ class BlogController extends Controller
 
     }
 
-    /**
-     * @return string
-     * @throws LoaderError
-     * @throws RuntimeError
-     * @throws SyntaxError
-     */
-    public function readAction()
-    {
-        $dataId = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT);
-        $articles = (new ArticleManager)->find($dataId);
-
-        $comments = (new CommentManager)->findAll($dataId);;
-
-
-        return $this->render('article.html.twig', compact('articles', 'comments'));
-
-
-    }
-
-
 }
-
-
-
-
-
-
-
-
-
